pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codewiki"
version = "0.1.0"
description = "AI-powered codebase wiki generator inspired by Karpathy's LLM Wiki pattern"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "litellm>=1.40.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "watchdog>=3.0.0",
    "jinja2>=3.1.0",
    "tiktoken>=0.5.0",
    "pathspec>=0.11.0",
    "tenacity>=8.2.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "ruff>=0.1.0",
]

[project.scripts]
codewiki = "codewiki.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/codewiki"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
